Re-factor coredump routines.  For each type of notes an output function is provided, which is used either to calculate the note size or output it to sbuf.  On the first pass the notes are registered in a list and the resulting size is found, on the second pass the list is traversed outputing notes to sbuf.  For the sbuf a drain routine is provided that writes data to a core file.
The main goal of the change is to make coredump to write notes
directly to the core file, without preliminary preparing them all in a
memory buffer.  Storing notes in memory is not a problem for the
current, rather small, set of notes we write to the core, but it may
becomes an issue when we start to store procstat notes.

Reviewed by:	jhb (initial version), kib
Discussed with:	jhb, kib
MFC after:	3 weeks


git-svn-id: svn://svn.freebsd.org/base/head@249486 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f