Fix the data corruption on the swap-backed md.
Assign the rv variable a success code if the pager was not asked for
the page.  Using an error code from the previous processed page caused
zeroing of the valid page, when e.g. the previous page was not
available in the pager.

Reported by:	lstewart
Sponsored by:	The FreeBSD Foundation
MFC after:	1 week


git-svn-id: svn://svn.freebsd.org/base/head@250966 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f